+++ Breakout/Hazards.h
@@ -3,4 +3,4 @@
 #include "Collisions.h"
 #include "sl.h"
 
-void blocks(Block block[blockRows][blockCols], int blueSprite, int& circlePosX, int& circlePosY, int radius, int& ballDirectionX, int& ballDirectionY);
+void blocks(Block block[blockRows][blockCols], int blueSprite, float& circlePosX, float& circlePosY, int radius, float& ballDirectionX, float& ballDirectionY);
